Swapping type to long to match other AAD number fields

+++ packages/kbn-alerts-as-data-utils/src/field_maps/alert_field_map.ts
@@ -114,7 +114,7 @@ export const alertFieldMap = {
     required: true,
   },
   [ALERT_RULE_REVISION]: {
-    type: 'integer',
+    type: 'long',
     array: false,
     required: true,
   },
